add js

+++ app/Http/Controllers/ComicController.php
@@ -42,17 +42,17 @@ class ComicController extends Controller
 
         $newComic = new Comic();
 
-            // $newComic->title = $data['title'];
-            // $newComic->description = $data['description'];
-            // $newComic->thumb = $data['description'];
-            // $newComic->price = $data['price'];
-            // $newComic->series = $data['series'];
-            // $newComic->sale_date = $data['sale_date'];
-            // $newComic->type = $data['type'];
-            $newComic->fill($data);
-            $newComic->save();
+        // $newComic->title = $data['title'];
+        // $newComic->description = $data['description'];
+        // $newComic->thumb = $data['description'];
+        // $newComic->price = $data['price'];
+        // $newComic->series = $data['series'];
+        // $newComic->sale_date = $data['sale_date'];
+        // $newComic->type = $data['type'];
+        $newComic->fill($data);
+        $newComic->save();
 
-            return redirect()->route('comics.index');
+        return redirect()->route('comics.index');
     }
 
     /**
@@ -61,9 +61,8 @@ class ComicController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Comic $comic)
     {
-        $comic = Comic::find($id);
         if ($comic) {
             return view('comic.show', compact('comic'));
         }
@@ -76,11 +75,10 @@ class ComicController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Comic $comic)
     {
-        $comic = Comic::find($id);
         if ($comic) {
-        return view('comic.edit', compact('comic'));
+            return view('comic.edit', compact('comic'));
         }
         abort(404);
     }
@@ -92,11 +90,10 @@ class ComicController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, Comic $comic)
     {
-        $comic = Comic::find($id);
 
-        if($comic) {
+        if ($comic) {
             $data = $request->all();
             $comic->update($data);
             $comic->save();
@@ -112,10 +109,12 @@ class ComicController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Comic $comic)
     {
-        $comic = Comic::find($id);
-        $comic->delete();
-        return redirect()->route('comics.index');
+        if ($comic) {
+            $comic->delete();
+            return redirect()->route('comics.index');
+        }
+        abort(404);
     }
 }
